Move all functions to separate module
